microservices/pred_producer/requirements.txt
kafka-python
torch
boto3
cassandra-driver
pandas
SQLAlchemy
sklearn